% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypefinder.R
\name{findPrototype}
\alias{findPrototype}
\title{Pickup Basketball Prototype Finder}
\usage{
findPrototype(relativeHeight, shooter, ballhandler, defense)
}
\arguments{
\item{relativeHeight}{from 1-5 how do you rank in height with your teammates}

\item{shooter}{logical, are you a shooter}

\item{ballhandler}{logical, can you dribble well}

\item{defense}{logical, are you a good defender}
}
\value{
player prototype
}
\description{
Pickup Basketball Prototype Finder
}
